Improve RSA parameter creation
This commit improves the way RSA parameter structures are created,
making it easier to create "types" of keys with only the required
parameters.
These changes were also propagated through to the transient key context,
allowing it to control the parameters and making it easier to implement
finer control over the keys in the future.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>